+++ plugin/src/net/groboclown/idea/p4ic/ui/config/props/RelativeConfigPartPanel.java
@@ -18,7 +18,6 @@ import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.text.StringUtil;
 import net.groboclown.idea.p4ic.P4Bundle;
-import net.groboclown.idea.p4ic.config.P4ProjectConfig;
 import net.groboclown.idea.p4ic.config.part.RelativeConfigCompositePart;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NotNull;
@@ -74,7 +73,7 @@ public class RelativeConfigPartPanel
 
     @Override
     public boolean isModified(@NotNull RelativeConfigCompositePart originalPart) {
-        return StringUtil.equals(originalPart.getName(), getConfigPart().getName());
+        return ! StringUtil.equals(originalPart.getName(), getConfigPart().getName());
     }
 
     {